+++ src/main/java/org/example/driverandfleetmanagementapp/repository/DriverRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.lang.NonNull;
-
 import java.util.List;
 import java.util.Optional;
 
-public interface DriverRepository extends JpaRepository<Driver, Integer> {
+
+public interface DriverRepository extends JpaRepository<Driver, Long> {
 
 
     //Solving N+1 problem with @EntityGraph
@@ -35,6 +35,6 @@ public interface DriverRepository extends JpaRepository<Driver, Integer> {
     @EntityGraph(attributePaths = {"vehicles"})
     @NonNull
     @Override
-    Optional<Driver> findById(@NonNull Integer id);
+    Optional<Driver> findById(@NonNull Long id);
 
 }
